// website/utils/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json }
  | Json[]

export interface Database {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          operationName: string
          query: string
          variables: Json
          extensions: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      pdf_element_locations: {
        Row: {
          element_idx: number
          element_type: string
          on_page: number
          pdf_id: string
        }
        Insert: {
          element_idx: number
          element_type: string
          on_page: number
          pdf_id: string
        }
        Update: {
          element_idx?: number
          element_type?: string
          on_page?: number
          pdf_id?: string
        }
      }
      pdf_processing_progress: {
        Row: {
          curr_step: number
          msg: string | null
          pdf_id: string
          success: boolean | null
          total_steps: number
        }
        Insert: {
          curr_step?: number
          msg?: string | null
          pdf_id: string
          success?: boolean | null
          total_steps: number
        }
        Update: {
          curr_step?: number
          msg?: string | null
          pdf_id?: string
          success?: boolean | null
          total_steps?: number
        }
      }
      pdf_summary: {
        Row: {
          pdf_id: string
          pdf_name: string
          pdf_summary: Json | null
        }
        Insert: {
          pdf_id: string
          pdf_name: string
          pdf_summary?: Json | null
        }
        Update: {
          pdf_id?: string
          pdf_name?: string
          pdf_summary?: Json | null
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
// ...
}